refactor: 인프라 계층 변경된 예외 디렉토리 적용
