+++ src/main/java/com/flab/just_10_minutes/payment/infrastructure/Iamport/IamportApiClient.java
@@ -11,7 +11,7 @@ import com.flab.just_10_minutes.payment.infrastructure.Iamport.response.IamportA
 import com.flab.just_10_minutes.payment.infrastructure.Iamport.response.IamportBillingCustomer;
 import com.flab.just_10_minutes.payment.infrastructure.Iamport.response.IamportResponse;
 import com.flab.just_10_minutes.payment.infrastructure.Iamport.response.IamportPayment;
-import com.flab.just_10_minutes.common.exception.iamport.IamportException;
+import com.flab.just_10_minutes.common.exception.infra.iamport.IamportException;
 import com.flab.just_10_minutes.common.infra.iamport.IamportConfig;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
